requirements.txt
amqp==5.0.6
asgiref==3.4.1
autopep8==1.5.7
backports.entry-points-selectable==1.1.0
billiard==3.6.4.0
celery==5.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.4
click==8.0.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
defusedxml==0.7.1
distlib==0.3.3
Django==3.2.7
django-crispy-forms==1.12.0
django-crontab==0.7.1
django-dotenv==1.4.2
django-environ==0.8.1
django-environment==0.3.8
django-filter==2.4.0
django-js-asset==1.2.2
django-mptt==0.13.2
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-silk==4.1.0
django-slick-reporting==0.6.0
django-templated-mail==1.1.1
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.20.0
environ==1.0
et-xmlfile==1.1.0
Faker==8.12.1
filelock==3.3.1
flower==1.0.0
gprof2dot==2021.2.21
humanize==3.12.0
idna==3.2
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.1
kombu==5.2.1
Markdown==3.3.4
MarkupSafe==2.0.1
numpy==1.21.3
oauthlib==3.1.1
openpyxl==3.0.9
packaging==21.0
pandas==1.3.4
pbr==5.6.0
Pillow==8.3.2
platformdirs==2.4.0
prometheus-client==0.12.0
prompt-toolkit==3.0.22
psycopg2-binary==2.9.1
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
PyJWT==2.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
redis==3.5.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.16
ruamel.yaml.clib==0.2.6
simplejson==3.17.5
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
stevedore==3.5.0
text-unidecode==1.3
toml==0.10.2
tornado==6.1
uritemplate==3.0.1
urllib3==1.26.6
vine==5.0.0
virtualenv==20.8.1
virtualenv-clone==0.5.7
virtualenvwrapper==4.8.4
wcwidth==0.2.5
